Refactor the scaffolding tool and add a reducer generator
