+++ frontend/scaffold.sh
@@ -1,0 +1,1 @@
+node ./src/scaffolding/scaffold.js
