+++ frontend/src/scaffolding/generators/index.js
@@ -1,0 +1,14 @@
+module.exports = [
+    {
+        display: 'General component',
+        generator: require('./component-generator')
+    },
+    {
+        display: 'Modal',
+        generator: require('./modal-generator')
+    },
+    {
+        display: 'Reducer',
+        generator: require('./reducer-generator')
+    }
+];
