+++ frontend/src/scaffolding/templates/component/[[name]].js
