+++ frontend/src/scaffolding/templates/component/[[name]].less
