+++ frontend/src/scaffolding/templates/modal/[[name]]-modal.js
