+++ frontend/src/scaffolding/templates/reducer/[[name]]-reducer.js
@@ -1,0 +1,18 @@
+import { createReducer } from 'utils/reducer-utils';
+
+// ------------------------------
+// Action Handlers
+// ------------------------------
+function onApplicationInit() {
+}
+
+// ------------------------------
+// Local util functions
+// ------------------------------
+
+// ------------------------------
+// Exported reducer function
+// ------------------------------
+export default createReducer({
+    APPLICATION_INIT: onApplicationInit
+});
